chore: clean up types to match design.
