+++ aegeri/process/src/lib.rs
@@ -4,4 +4,3 @@ pub use aegeri_message;
 pub use gossamer;
 pub use gwrdfa_resample;
 pub use parabyzantine;
-pub use task::{AegeriTaskFlow, TaskFlowError, TaskFlowResult};
